move ast to dml conversion into own module
