+++ libs/datamodel/core/src/validator/mod.rs
@@ -1,15 +1,3 @@
 pub mod ast_to_dml;
-mod common;
 mod directive;
 pub mod dml_to_ast;
-mod invalid_model_names;
-mod lift;
-mod precheck;
-mod standardise;
-mod validate;
-mod validation_pipeline;
-
-use lift::*;
-use standardise::*;
-use validate::*;
-pub use validation_pipeline::*;
